Move input elements into their own subdirectory
