+++ src/app/components/elements/UserEmailPreferences.tsx
@@ -1,7 +1,7 @@
 import {CardBody, Col, FormGroup, Row, default as RS, Table} from "reactstrap";
 import React, {useMemo} from "react";
 import {UserEmailPreferences} from "../../../IsaacAppTypes";
-import {TrueFalseRadioInput} from "./TrueFalseRadioInput";
+import {TrueFalseRadioInput} from "./inputs/TrueFalseRadioInput";
 import {useSelector} from "react-redux";
 import {AppState} from "../../state/reducers";
 import {validateEmailPreferences} from "../../services/validation";
