+++ src/app/components/elements/inputs/AnonUserExamBoardPicker.tsx
@@ -1,10 +1,10 @@
 import React, {ChangeEvent} from "react";
 import {FormGroup, Input, Label} from "reactstrap";
-import {determineExamBoardFrom} from "../../services/examBoard";
-import {EXAM_BOARD} from "../../services/constants";
+import {determineExamBoardFrom} from "../../../services/examBoard";
+import {EXAM_BOARD} from "../../../services/constants";
 import {useDispatch, useSelector} from "react-redux";
-import {setAnonUserPreferences} from "../../state/actions";
-import {AppState} from "../../state/reducers";
+import {setAnonUserPreferences} from "../../../state/actions";
+import {AppState} from "../../../state/reducers";
 
 interface AnonUserExamBoardPicker {className?: string; hideLabel?: boolean}
